refactor(platform): extract PlatformBaseUserManager (audit H3)
The lockout-aware ``authenticate``, ``authenticate_password``, and
HIBP-validating ``validate_password`` had drifted between MBK and MJH
despite being structurally identical. This PR extracts them into
``platform_shared.auth.user_manager.PlatformBaseUserManager`` and has
both apps subclass it. ~190 LOC of duplicated security logic moves to
one canonical implementation.

What stays in the apps:
- ``on_after_register`` — MBK creates org+member+tax_profile, MJH just
  triggers verification email. Tier 3 by definition.
- ``on_after_request_verify`` — each app uses its own email-sender module.
- ``on_after_forgot_password`` — same.
- ``on_after_reset_password`` — only MBK has the TOTP-on-reset escape hatch.
- ``on_after_update`` — only MBK logs PASSWORD_CHANGE_SUCCESS.
- ``on_after_verify`` — different log shapes per app.

Configuration is exposed via @property in each subclass so tests that
monkeypatch ``app.core.auth.settings.hibp_enabled`` (or other fields)
keep taking effect on the next request — the property re-reads at call
time. The shared base provides sane class-attribute defaults for tests
that construct managers without a settings object.

MJH behavior change: a request to ``/auth/jwt/login`` for an unverified
user now returns ``LOGIN_BAD_CREDENTIALS`` (manager rejects pre-router)
rather than ``LOGIN_USER_NOT_VERIFIED``. The MJH frontend always uses
``/auth/totp/login`` as its primary login path (which still surfaces
``LOGIN_USER_NOT_VERIFIED`` correctly), so this only affects direct API
hits to /auth/jwt/login — none today. Test asserts updated.

Test patches updated for symbols that moved out of ``app.core.auth``:
- ``app.core.auth.is_password_pwned`` →
  ``platform_shared.auth.user_manager.is_password_pwned``
- ``app.core.auth.emit_locked_login_event`` →
  ``platform_shared.auth.user_manager.emit_locked_login_event``
- ``UserManager.__bases__[1].authenticate`` →
  ``fastapi_users.BaseUserManager.authenticate`` (direct import, no MRO indexing)

Verified: 197 MBK + 64 MJH auth-related tests pass; 425 shared tests pass.
